Refactor login into a single file
