+++ src/components/account/LogIn.tsx
@@ -1,16 +1,71 @@
-import { Message } from "../common/Message";
-import { defaultLoginFormFieldsState } from "./constants/defaultState";
-import { useSignIn } from "./hooks/useLogIn";
-import { loginSchema } from "./schema";
-import { LogInFormFieldsType } from "./types";
+import useFirebase from "../../hooks/useFirebase";
+import { AsyncActionStateType } from "../../types/asyncActionStateType";
+import { Message } from "./common/Message";
 import { yupResolver } from "@hookform/resolvers/yup";
-import { FC, Fragment } from "react";
+import { signInWithEmailAndPassword } from "firebase/auth";
+import { FC, Fragment, useCallback, useState } from "react";
 import { useForm } from "react-hook-form";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
+import { object, string } from "yup";
+
+type LogInFormFieldsType = {
+  email: string;
+  password: string;
+};
+
+const loginSchema = object()
+  .shape({
+    email: string().required().email("Invalid email address"),
+    password: string().required().min(8, "Invalid password"),
+  })
+  .required();
+
+const messageByErrorCodeMap: Record<string, string> = {
+  "auth/user-not-found": "Couldn't find your account",
+  "auth/wrong-password":
+    "Unable to log in, please be sure to use the proper email/password",
+};
+
+type LogInActionsType = {
+  logIn: (data: LogInFormFieldsType) => void;
+};
+
+export const useSignIn = (): [AsyncActionStateType, LogInActionsType] => {
+  const firebase = useFirebase();
+  const navigate = useNavigate();
+  const [state, setState] = useState<AsyncActionStateType>({
+    status: "idle",
+    message: "",
+  });
+
+  const logIn = useCallback(
+    (data: LogInFormFieldsType) => {
+      signInWithEmailAndPassword(firebase.auth, data.email, data.password)
+        .then(() => {
+          setState({ status: "success", message: "Successfully logged in" });
+          navigate("/tos");
+        })
+        .catch((error) => {
+          setState({
+            status: "failure",
+            message:
+              messageByErrorCodeMap[error.code] ??
+              "There was an error loggin in to your account, please try again",
+          });
+        });
+    },
+    [firebase.auth, navigate],
+  );
+
+  return [state, { logIn }];
+};
 
 const LogIn: FC = () => {
   const form = useForm<LogInFormFieldsType>({
-    defaultValues: defaultLoginFormFieldsState,
+    defaultValues: {
+      email: "",
+      password: "",
+    },
     resolver: yupResolver(loginSchema),
   });
