--- src/components/account/LogIn/constants/defaultState.ts
@@ -1,6 +1,0 @@
-import { LogInFormFieldsType } from "../types";
-
-export const defaultLoginFormFieldsState: LogInFormFieldsType = {
-  email: "",
-  password: "",
-};
