--- src/components/account/LogIn/hooks/useLogIn/types.ts
@@ -1,5 +1,0 @@
-import { LogInFormFieldsType } from "../../types";
-
-export type LogInActionsType = {
-  logIn: (data: LogInFormFieldsType) => void;
-};
