--- src/components/account/LogIn/types.ts
@@ -1,4 +1,0 @@
-export type LogInFormFieldsType = {
-  email: string;
-  password: string;
-};
